Add a classes for proxying actor messages across SQS between services
When splitting out background processing, it became apparent that we needed to have a way
to send messages from an API service to actors running in the background processing jobs service.

With minimal amount of refactoring required, we can do this by binding a proxy actor to the original
actor name in the API service, which send all messages it gets over SQS to a receiver in the jobs service.
The receiver will forward all messages received on its SQS queue to the original actor that now lives in
the jobs service.